Updated Data Ingestion and Requirements to check for already installed data and libraries before installing them

+++ Data_Ingestion.R
@@ -1,13 +1,19 @@
 download_data <- function() {
-  kgl_auth(creds_file = 'kaggle.json')
-  
-  response <- kgl_datasets_download_all(owner_dataset = "kartik2112/fraud-detection")
-  
-  download.file(response[["url"]], "archive.zip", mode="wb")
-  
-  unzip_result <- unzip("archive.zip", exdir = "data/", overwrite = TRUE)
-  
-  unlink("archive.zip")
+  if(!(file.exists("data"))) {
+    
+    kgl_auth(creds_file = 'kaggle.json')
+    
+    response <- kgl_datasets_download_all(owner_dataset = "kartik2112/fraud-detection")
+    
+    download.file(response[["url"]], "archive.zip", mode="wb")
+    
+    unzip_result <- unzip("archive.zip", exdir = "data/", overwrite = TRUE)
+    
+    unlink("archive.zip")
+  }
+  else{
+    print("Data Already exists")
+  }
 }
